refactor: share Photographer and Video primitives across the data types
VideoItem (works) and LiveVideo (live) were byte-identical, and the photographer
credit was modelled three ways (a shared Photographer type plus two inline
{name,url?} objects). Extract a types/media.ts with Photographer and Video; works
and live use the shared Video, works Image and AboutImage use the shared
Photographer, and LightboxVideo extends Video. Type-only, output unchanged.
