+++ src/types/index.ts
@@ -2,6 +2,7 @@ export * from './about';
 export * from './contact';
 export * from './lightbox';
 export * from './live';
+export * from './media';
 export * from './navigation';
 export * from './press';
 export * from './schema';
